+++ pkg/analysis_server_plugin/lib/edit/fix/dart_fix_context.dart
@@ -6,7 +6,7 @@ import 'package:analysis_server_plugin/edit/fix/fix_context.dart';
 import 'package:analysis_server_plugin/src/correction/change_workspace.dart';
 import 'package:analyzer/dart/analysis/results.dart';
 import 'package:analyzer/dart/element/element.dart';
-import 'package:analyzer/error/error.dart';
+import 'package:analyzer/diagnostic/diagnostic.dart';
 import 'package:analyzer/instrumentation/service.dart';
 import 'package:analyzer/src/dart/analysis/driver_based_analysis_context.dart';
 import 'package:analyzer/src/dart/analysis/file_state_filter.dart';
@@ -49,7 +49,7 @@ class DartFixContext implements FixContext {
       _cachedTopLevelDeclarations = {};
 
   @override
-  final AnalysisError error;
+  final Diagnostic error;
 
   DartFixContext({
     required this.instrumentationService,