+++ pkg/analysis_server_plugin/lib/edit/fix/fix_context.dart
@@ -2,7 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:analyzer/error/error.dart';
+import 'package:analyzer/diagnostic/diagnostic.dart';
 import 'package:analyzer_plugin/utilities/fixes/fixes.dart';
 
 /// Provides context information for [FixContributor]s.
@@ -10,5 +10,5 @@ import 'package:analyzer_plugin/utilities/fixes/fixes.dart';
 /// Clients may not extend, implement or mix-in this class.
 abstract interface class FixContext {
   /// The error to fix.
-  AnalysisError get error;
+  Diagnostic get error;
 }