+++ sdk/lib/_internal/js_dev_runtime/patch/internal_patch.dart
@@ -113,3 +113,11 @@ Future<Object?> loadDynamicModule({Uri? uri, Uint8List? bytes}) {
   }
   return completer.future;
 }
+
+@patch
+@pragma("vm:entry-point")
+abstract interface class IsolateGroup {
+  @patch
+  static Object _runSync(Object computation) =>
+      throw UnsupportedError("_runSync");
+}